+++ objectbox_inspector/lib/src/objectbox_inspector_page.dart
@@ -24,7 +24,7 @@ import 'models/inspectable_box.dart';
 ///   ),
 /// );
 /// ```
-class ObjectboxInspectorPage extends StatelessWidget {
+class ObjectboxInspectorPage extends StatefulWidget {
   final List<InspectableBox> boxes;
 
   const ObjectboxInspectorPage({
@@ -33,39 +33,111 @@ class ObjectboxInspectorPage extends StatelessWidget {
   });
 
   @override
+  State<ObjectboxInspectorPage> createState() => _ObjectboxInspectorPageState();
+}
+
+class _ObjectboxInspectorPageState extends State<ObjectboxInspectorPage> {
+  final searchController = TextEditingController();
+  var boxes = <InspectableBox>[];
+  bool isSearching = false;
+  String? searchQuery;
+
+  @override
+  void initState() {
+    super.initState();
+    boxes = widget.boxes;
+    searchController.addListener(() {
+      setState(() {
+        searchQuery = searchController.text.toLowerCase();
+        if (searchQuery != null && searchQuery!.isNotEmpty) {
+          boxes = widget.boxes
+              .where(
+                (box) => box.boxName.toLowerCase().contains(searchQuery!),
+              )
+              .toList();
+        } else {
+          boxes = widget.boxes;
+        }
+      });
+    });
+  }
+
+  @override
+  void dispose() {
+    searchController.dispose();
+    super.dispose();
+  }
+
+  @override
   Widget build(BuildContext context) {
     final mq = MediaQuery.of(context);
-    final tt = Theme.of(context).textTheme;
     final cs = Theme.of(context).colorScheme;
 
     return Scaffold(
-      appBar: AppBar(
-        title: Text(
-          'Objectbox Inspector',
-          style: tt.titleLarge?.copyWith(
-            fontWeight: FontWeight.w500,
+      body: CustomScrollView(
+        keyboardDismissBehavior: ScrollViewKeyboardDismissBehavior.onDrag,
+        slivers: [
+          SliverAppBar(
+            title: isSearching
+                ? TextFormField(
+                    controller: searchController,
+                    decoration: InputDecoration(
+                        hintText: 'Name contains...',
+                        border: OutlineInputBorder(
+                          borderSide: BorderSide(
+                            color: cs.outline,
+                          ),
+                          borderRadius: BorderRadius.circular(12),
+                        ),
+                        contentPadding: const EdgeInsets.fromLTRB(10, 8, 10, 8),
+                        isDense: true),
+                  )
+                : const Text('ObjectBox Inspector'),
+            titleSpacing: 0,
+            centerTitle: true,
+            backgroundColor: cs.surface,
+            pinned: true,
+            actions: [
+              IconButton(
+                onPressed: () {
+                  setState(() {
+                    isSearching = !isSearching;
+                    if (!isSearching) {
+                      searchController.clear();
+                    }
+                  });
+                },
+                icon: Icon(
+                  isSearching ? Icons.close : Icons.search,
+                ),
+              ),
+            ],
           ),
-        ),
-        backgroundColor: cs.surface,
-      ),
-      body: ListView.builder(
-        padding: EdgeInsets.only(
-          bottom: mq.viewPadding.bottom + 20,
-        ),
-        itemCount: boxes.length,
-        itemBuilder: (context, index) {
-          final box = boxes[index];
+          SliverList.builder(
+            itemCount: boxes.length,
+            itemBuilder: (context, index) {
+              final box = boxes[index];
 
-          return BoxListTile(
-            box: box,
-            onTap: () {
-              BoxPageNavigator.push(
-                context,
-                box.boxName,
+              return BoxListTile(
+                box: box,
+                highlight: searchQuery,
+                onTap: () {
+                  BoxPageNavigator.push(
+                    context,
+                    box.boxName,
+                  );
+                },
               );
             },
-          );
-        },
+          ),
+          SliverToBoxAdapter(
+            child: Padding(
+              padding: EdgeInsets.only(
+                bottom: mq.viewPadding.bottom + 20,
+              ),
+            ),
+          ),
+        ],
       ),
     );
   }